Adicionando findAll e findById em Pessoa

+++ src/main/java/br/ufrn/imd/utravel/repository/GenericRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 public interface GenericRepository <T extends AbstractModel> {
     List<T> findAll();
 
-    Optional<T> findById(Long id);
+    Optional<T> findById(Integer id);
 
     T save(T modelo);
